move lbx code to its own package
